requirements.txt
dnspython==2.7.0
et_xmlfile==2.0.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.11
pymongo==4.16.0
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
schedule==1.2.2
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
tzdata==2025.3
